% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_complex_area}
\alias{get_complex_area}
\title{Get Hawaii Complex Area from school}
\usage{
get_complex_area(school_code)
}
\arguments{
\item{school_code}{Character vector of school codes}
}
\value{
Character vector of Complex Area names
}
\description{
Maps schools to their Complex Area (geographic grouping).
Hawaii has 15 Complex Areas organized around high schools.
}
\keyword{internal}
