% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_hidoe_url}
\alias{get_hidoe_url}
\title{Get HIDOE enrollment URL for a specific year}
\usage{
get_hidoe_url(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
URL string
}
\description{
Returns the URL for downloading HIDOE enrollment data.
}
\keyword{internal}
