% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{standardize_hidoe_columns}
\alias{standardize_hidoe_columns}
\title{Standardize HIDOE column names}
\usage{
standardize_hidoe_columns(df)
}
\arguments{
\item{df}{Data frame with raw HIDOE columns}
}
\value{
Data frame with standardized column names
}
\description{
Maps HIDOE column names to standardized names.
}
\keyword{internal}
