% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_dbedt_enrollment}
\alias{download_dbedt_enrollment}
\title{Download DBEDT State Data Book enrollment data (2010-2017)}
\usage{
download_dbedt_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end (2010-2017)}
}
\value{
Data frame with enrollment data
}
\description{
Downloads enrollment data from the Hawaii DBEDT State Data Book.
Files are Excel tables with state-level and county-level enrollment.
}
\keyword{internal}
